Split hotspot renderer test slices
