+++ crates/pixel_world_bridge/src/render.rs
@@ -11,6 +11,10 @@ use fragment_visuals::reconcile_fragments;
 mod agent_silhouette;
 use agent_silhouette::{PixelWorldAgentSilhouetteVisual, reconcile_agent_silhouettes};
 
+#[path = "render_hotspot_core.rs"]
+mod hotspot_core;
+use hotspot_core::{PixelWorldHotspotCoreVisual, reconcile_hotspot_cores};
+
 const LOCATION_HIT_HALF_SIZE: f64 = 8.0;
 const AGENT_HIT_HALF_SIZE: f64 = 8.0;
 const FRAGMENT_HIDDEN_THRESHOLD_PX: f64 = 1.5;
@@ -31,11 +35,6 @@ const AGENT_LAYER_Z: f32 = 3.0;
 const AGENT_CORE_LAYER_Z_OFFSET: f32 = 0.01;
 const AGENT_CORE_SIZE_SCALE: f32 = 0.32;
 const AGENT_CORE_COLOR: Color = Color::srgba_u8(224, 242, 254, 238);
-const HOTSPOT_CORE_LAYER_Z_OFFSET: f32 = 0.01;
-const HOTSPOT_CORE_SIZE_SCALE: f64 = 0.30;
-const HOTSPOT_CORE_MIN_SIZE_PX: f64 = 2.0;
-const HOTSPOT_CORE_MAX_SIZE_PX: f64 = 5.0;
-const HOTSPOT_CORE_COLOR: Color = Color::srgba_u8(226, 232, 240, 230);
 const SELECTED_ENTITY_LAYER_Z_OFFSET: f32 = 1.0;
 const SELECTED_ENTITY_SIZE_SCALE: f64 = 1.35;
 const SELECTED_LOCATION_CUE_THICKNESS_PX: f32 = 2.0;
@@ -144,13 +143,6 @@ pub(crate) struct PixelWorldLinkVisual {
 
 #[derive(Component)]
 pub(crate) struct PixelWorldHotspotVisual {
-    id: String,
-}
-
-/// A non-interactive neutral inset that makes every hotspot read as a layered
-/// signal without assigning an additional semantic color to its kind.
-#[derive(Component)]
-struct PixelWorldHotspotCoreVisual {
     id: String,
 }
 
@@ -977,74 +969,6 @@ fn reconcile_hotspots(
     }
 
     despawn_stale_entities(commands, &mut runtime.hotspot_entities, &active_ids);
-}
-
-fn reconcile_hotspot_cores(
-    commands: &mut Commands,
-    runtime: &BevyRuntimeState,
-    existing_cores: &Query<(Entity, &PixelWorldHotspotCoreVisual)>,
-    width: f64,
-    height: f64,
-    animation_ms: f64,
-) {
-    let mut existing_by_id = HashMap::new();
-    for (entity, core) in existing_cores.iter() {
-        existing_by_id.insert(core.id.clone(), entity);
-    }
-
-    let Some(render_state) = runtime.render_state.as_ref() else {
-        for entity in existing_by_id.into_values() {
-            commands.entity(entity).despawn();
-        }
-        return;
-    };
-    let Some(world_bounds) = render_state.world_bounds.as_ref() else {
-        for entity in existing_by_id.into_values() {
-            commands.entity(entity).despawn();
-        }
-        return;
-    };
-
-    let mut active_ids = HashSet::new();
-    for (index, hotspot) in render_state.visual_hotspots.iter().enumerate() {
-        let Some((canvas_x, canvas_y)) =
-            to_canvas_point(&hotspot.pos, world_bounds, width, height, &runtime.camera)
-        else {
-            continue;
-        };
-        active_ids.insert(hotspot.id.clone());
-        let pulse = 1.0 + (0.1 * ((animation_ms / 280.0) + index as f64).sin());
-        let base_size = hotspot.size_hint_px.unwrap_or(10.0) * pulse;
-        let core_size = (base_size * HOTSPOT_CORE_SIZE_SCALE)
-            .clamp(HOTSPOT_CORE_MIN_SIZE_PX, HOTSPOT_CORE_MAX_SIZE_PX)
-            as f32;
-        let sprite = sprite_for_square(HOTSPOT_CORE_COLOR, core_size);
-        let transform = Transform::from_translation(to_bevy_translation(
-            canvas_x,
-            canvas_y,
-            width,
-            height,
-            1.5 + HOTSPOT_CORE_LAYER_Z_OFFSET,
-        ));
-
-        if let Some(entity) = existing_by_id.get(&hotspot.id).copied() {
-            commands.entity(entity).insert((sprite, transform));
-        } else {
-            commands.spawn((
-                sprite,
-                transform,
-                PixelWorldHotspotCoreVisual {
-                    id: hotspot.id.clone(),
-                },
-            ));
-        }
-    }
-
-    for (id, entity) in existing_by_id {
-        if !active_ids.contains(&id) {
-            commands.entity(entity).despawn();
-        }
-    }
 }
 
 fn clear_runtime_visuals(commands: &mut Commands, runtime: &mut BevyRuntimeState) {
